Updates to allow for constraints to be updated
The previous implementation only allows for constraints to be created
upon calls to ToLayoutConstraints. The updated implementation allows
updates to updatable properties, so that constraints can be updated
later on. This is super helpful when looking to build interfaces that
need resizing and where having animation of those UIs would be helpful.