Add tvOS as target for FluentLayouts
Xamarin.tvOS is using a different project type GUID so it was not possible to use different build configurations to switch between the targets. Instead, I created a new project for tvOS and a shared project where I moved all the existing code to. This way we can easily build two projects, get two DLL files and still share the code.